+++ apps/web/utils/Constants.ts
@@ -26,6 +26,10 @@ export type PasswordVisibilityKey =
 export const STRING = "string";
 export const VIEW = "view";
 export const HTTP_STATUS_UNAUTHORIZED = 401;
+export const WEBSITE = "website";
+export const OPEN_GRAPH_LOCALE_EN_US = "en_US";
+export const TWITTER_CARD_SUMMARY_LARGE_IMAGE = "summary_large_image";
+export const KIIBEE_LANDING_PAGE_PREVIEW_ALT = "Kiibee landing page preview";
 
 export const CARD_BRANDS = {
   VISA: "visa",
